Fix patching. Move default patching to __init__
